+++ module/schemas/actorCommonSchema.js
@@ -1,0 +1,3 @@
+import { ARM5E } from "../config.js";
+
+const fields = foundry.data.fields;